requirements.txt
flask
flask_sqlalchemy
requests
pyyaml
pillow
